+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 
 ## 1.0.0
 
-Initial release of Insert Multiple Rows extension
+- Initial release of Insert Multiple Rows extension
+
+## 2.0.0
+
+- Rename "Insert hexadecimal command" to "Insert bitfield command"
+- Add padding characoter setting to "Insert decimal command"
+- Update user input filter of "Insert bitfield command"